cli/command/container: don't mutate ConfigFile.DetachKeys
This code was introduced in https://github.com/moby/moby/commit/15aa2a663b47b6126a66efefcadb64edfbffb9f5,
but from those changes, it appears that overwriting the config value was
merely out of convenience, and that struct being used as an intermediate.

While changing the config here should be mostly ephemeral, and not written
back to the config-file, let's be clear on intent, and not mutatte the config
as part of this code.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>