info: remove printSecurityOptionsWarnings, printServerWarnings
Docker Engine 1.13 (API v1.25) added an option to set a custom default seccomp
profile on the daemon (see [moby/moby@b237189]). A warning was added on the
client-side if a non-default profile was set.

Docker Engine 23.0 (API v1.42) added warnings about non-default seccomp
profiles to the "info" response ([moby/moby@04f932a]), and the client was
updated to skip generating client-side warnings for API v1.42 and up in
[docker/cli@8964595].

These warnings are purely informational, and given that Docker Engine versions
before 23.0 have reached EOL, and any current version of the Engine now returns
the Warnings, it should be safe to remove the client-side fall back logic.

This patch removes the client-side fall back code for warnings that was
added in 8964595692766444fb84445931889677af5f55eb.

[moby/moby@b237189]: https://github.com/moby/moby/commit/b237189e6c8a4f97be59f08c63cdcb1f2f4680a8
[moby/moby@04f932a]: https://github.com/moby/moby/commit/04f932ac862c05a5af2a5950b5e1744100b14eea
[docker/cli@8964595]: https://github.com/docker/cli/commit/8964595692766444fb84445931889677af5f55eb

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>